autentikasi

+++ routes/web.php
@@ -1,19 +1,27 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ResidentController;
 
-Route::get('/', function () {
-    return view('layout/app');
-});
+// Route::get('/', function () {
+//     return view('welcome');
+// });
+
+// auth
+Route::get('/', [AuthController::class, 'login']);
+Route::post('/login', [AuthController::class, 'authenticate']);
+Route::post('/logout', [AuthController::class, 'logout']);
+Route::get('/register', [AuthController::class, 'registerView']);
+Route::post('/register', [AuthController::class, 'register']);
 
 Route::get('/dashboard', function () {
     return view('pages.dashboard');
-});
+})->middleware('role:Admin,User');
 
-Route::get('/resident', [ResidentController::class, 'index']);
-Route::get('/resident/create', [ResidentController::class, 'create']);
-Route::get('/resident/{id}', [ResidentController::class, 'edit']);
-Route::post('/resident', [ResidentController::class, 'store']);
-Route::put('/resident/{id}', [ResidentController::class, 'update']);
-Route::delete('/resident/{id}', [ResidentController::class, 'destroy']);
+Route::get('/resident', [ResidentController::class, 'index'])->middleware('role:Admin');
+Route::get('/resident/create', [ResidentController::class, 'create'])->middleware('role:Admin');
+Route::get('/resident/{id}', [ResidentController::class, 'edit'])->middleware('role:Admin');
+Route::post('/resident', [ResidentController::class, 'store'])->middleware('role:Admin');
+Route::put('/resident/{id}', [ResidentController::class, 'update'])->middleware('role:Admin');
+Route::delete('/resident/{id}', [ResidentController::class, 'destroy'])->middleware('role:Admin');
